Refactor: 소켓 폴더 생성
소켓 폴더에서 소켓 관련 처리
